# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Anti-Scam Trainer API
  version: 0.1.0
  description: |
    Публичный API интерактивного антискам-тренажёра.

    Реализованная часть: проверка живости и готовности, каталог учебных
    сценариев, запуск попытки, восстановление её состояния, отправка выбора
    прогресс пользователя и еженедельный тест, создаваемый через Groq или Grok.

    ### Анонимный профиль

    Учётных записей нет. При первом обращении к `/api/v1/**` сервер выдаёт
    cookie `ast_profile` с анонимным идентификатором профиля:

    - `HttpOnly`, `SameSite=Lax` (настраивается), `Path=/`, длительный `Max-Age`;
    - `Secure` включается в публичном развёртывании по HTTPS;
    - при отсутствии или неверном формате cookie сервер выдаёт новый случайный
      идентификатор; валидная cookie является bearer-секретом доступа к истории;
    - к профилю привязываются попытки, поэтому браузер обязан отправлять
      cookie обратно: браузерный клиент выполняет каждый запрос с
      `credentials: "include"`.

    Попытка принадлежит выдавшему её профилю. Обращение к чужой попытке
    возвращает `403 ATTEMPT_FORBIDDEN`, даже если идентификатор известен.
    Потеря cookie означает потерю доступа к истории: восстановления нет.

    ### Доступ из браузера

    Рекомендуемая схема — общий origin: фронтенд и API отдаются одним адресом
    (в Compose это nginx фронтенда, проксирующий `/api`), тогда cookie
    остаётся same-site и межсайтовый доступ не нужен.

...

...

...
  license:
    name: MIT
...

...
security: []
servers:
...
paths:
...
  /api/v1/attempts/{attemptId}/choices:
    post:
      tags: [attempts]
      summary: Отправить выбор
      description: |
        Применяет выбор пользователя и продвигает прохождение до следующего
        решения или до финала.

        Сервер проверяет запрос по своим данным: попытка не завершена, узел
        совпадает с текущим, узел действительно является точкой решения,
        а вариант принадлежит этому узлу зафиксированной версии сценария.

        Влияние выбора на результат и навыки берётся только из сценария —
        клиент не может прислать ни очки, ни эффекты.

        ### Повторная отправка

        `idempotencyKey` защищает от двойного применения при обрыве связи:

        - тот же ключ с теми же `nodeId` и `choiceId` возвращает **тот же**
          ответ, эффекты повторно не применяются;
        - тот же ключ с другими данными — `409 IDEMPOTENCY_KEY_CONFLICT`;
        - ключ действует в пределах одной попытки.

        Клиент должен генерировать ключ сам и повторять запрос с тем же
        значением, а не с новым.
      operationId: submitChoice
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitChoiceRequest"
            example:
              nodeId: channel-decision
              choiceId: stay-on-platform
              idempotencyKey: 0f8c2a1b-7d4e-4a19-9c3f-2b6e5d1a8c47
      responses:
        "200":
          description: Выбор применён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transition"
        "400":
          description: Тело запроса некорректно или не заполнены обязательные поля
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: INVALID_REQUEST
                  message: Поле idempotencyKey обязательно.
                  requestId: 7b2c1a9e4f
        "403":
          $ref: "#/components/responses/AttemptForbidden"
        "404":
          description: Попытка не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: ATTEMPT_NOT_FOUND
                  message: Попытка не найдена.
                  requestId: 7b2c1a9e4f
        "409":
          description: |
            Состояние попытки не позволяет применить выбор:

            - `ATTEMPT_ALREADY_COMPLETED` — прохождение завершено;
            - `STALE_NODE` — прислан не тот узел, на котором стоит попытка;
            - `IDEMPOTENCY_KEY_CONFLICT` — ключ повтора уже использован с другими данными;
            - `CONCURRENT_TRANSITION` — попытку изменил другой запрос.

            В каждом случае нужно перечитать состояние попытки.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: STALE_NODE
                  message: Присланный узел не совпадает с текущим. Перечитайте состояние попытки.
                  requestId: 7b2c1a9e4f
        "422":
          description: |
            Выбор не применим к текущему узлу:

            - `NODE_NOT_DECISION` — на текущем узле выбор не принимается;
            - `CHOICE_NOT_FOUND` — вариант не принадлежит текущему узлу.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: CHOICE_NOT_FOUND
                  message: Такой вариант выбора недоступен на текущем узле.
                  requestId: 7b2c1a9e4f
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  responses:
    AttemptForbidden:
      description: |
        Попытка принадлежит другому анонимному профилю.

        Обычная причина — потерянная или незаданная cookie `ast_profile`:
        браузер обратился к попытке от имени нового профиля.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: ATTEMPT_FORBIDDEN
              message: Попытка принадлежит другому пользователю.
              requestId: 7b2c1a9e4f
    PayloadTooLarge:
      description: Тело запроса превышает допустимый размер
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: PAYLOAD_TOO_LARGE
              message: Тело запроса превышает допустимый размер.
              requestId: 7b2c1a9e4f
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: Внутренняя ошибка сервера.
              requestId: 7b2c1a9e4f
  schemas:
...
    AttemptStatus:
      type: string
      description: |
        Состояние прохождения. Попытка становится `completed` только после
        того, как сервер довёл её до финального узла.
      enum: [in_progress, completed]
    NodeType:
      type: string
      description: |
        Тип раскрытого узла: реплика диалога, точка принятия решения или финал.
      enum: [message, decision, terminal]
    OutcomeType:
      type: string
      description: Характер финала сценария.
      enum: [safe, unsafe]
    Severity:
      type: string
      description: Насколько опасным оказался сделанный выбор.
      enum: [safe, warning, dangerous]
    AttemptChoice:
      type: object
      description: |
        Доступное действие. Публикуются только идентификатор и подпись:
        следующий узел и влияние выбора остаются на сервере.
      required: [id, label]
      properties:
        id:
          type: string
          example: stay-on-platform
        label:
          type: string
          example: Остаться в чате площадки
    AttemptOutcome:
      type: object
      description: Итог сценария. Появляется только у финального узла.
      required: [type, title, explanation]
      properties:
        type:
          $ref: "#/components/schemas/OutcomeType"
        title:
          type: string
        explanation:
          type: string
    AttemptNode:
      type: object
      description: |
        Раскрытый узел сценария. Набор заполненных полей зависит от типа:
        `message` содержит отправителя и текст, `decision` — вопрос и варианты,
        `terminal` — итог.
      required: [id, type]
      properties:
        id:
          type: string
          example: channel-decision
        type:
          $ref: "#/components/schemas/NodeType"
        sender:
          type: string
          enum: [buyer, seller, system]
          description: |
            Кто прислал сообщение. Только для узлов типа `message`.
            Роль игрока здесь не встречается: его слова приходят полем
            `playerReply` сделанного решения. `system` — не участник сделки,
            а факт обстановки, поэтому в ленте переписки ему не место.
          example: seller
        text:
          type: string
          description: Текст реплики. Только для узлов типа `message`.
        prompt:
          type: string
          description: Вопрос пользователю. Только для узлов типа `decision`.
        choices:
          type: array
          description: Доступные действия. Только для узлов типа `decision`.
          items:
            $ref: "#/components/schemas/AttemptChoice"
        outcome:
          $ref: "#/components/schemas/AttemptOutcome"
    AttemptConsequence:
      type: object
      description: |
        Учебное объяснение последствий выбора. Появляется только после того,
        как выбор сделан.
      required: [severity, title, explanation]
      properties:
        severity:
          $ref: "#/components/schemas/Severity"
        title:
          type: string
        explanation:
          type: string
        realWorldRule:
          type: string
          description: Короткое правило, применимое в реальной сделке.
...
    SubmitChoiceRequest:
      type: object
      description: |
        Запрос содержит только то, что сервер не может вычислить сам.
        Следующего узла, изменения результата и эффектов здесь нет намеренно.
      required: [nodeId, choiceId, idempotencyKey]
      properties:
        nodeId:
          type: string
          description: |
            Узел, который видел пользователь. Служит для сверки с текущим узлом
            попытки: расхождение даёт `409 STALE_NODE`.
          example: channel-decision
        choiceId:
          type: string
          description: Выбранный вариант из числа показанных на этом узле.
          example: stay-on-platform
        idempotencyKey:
          type: string
          description: |
            Уникальное значение, которое генерирует клиент. При повторе того же
            запроса используйте тот же ключ, чтобы эффекты не применились дважды.
          example: 0f8c2a1b-7d4e-4a19-9c3f-2b6e5d1a8c47
    AcceptedChoice:
      type: object
      description: Выбор, который сервер принял и применил.
      required: [nodeId, choiceId, label, playerReply]
      properties:
        nodeId:
          type: string
        choiceId:
          type: string
        label:
          type: string
        playerReply:
          type: string
          description: Реплика игрока, которая добавляется в ленту переписки.
    Transition:
      type: object
      description: |
        Результат одного шага прохождения: что было принято, чему это учит и
        что показать дальше. Содержит только новые узлы — уже показанные
        клиент получил раньше.
      required:
        [attemptId, status, score, acceptedChoice, consequence, revealedNodes, currentNodeId]
      properties:
        attemptId:
          type: string
        status:
          $ref: "#/components/schemas/AttemptStatus"
        score:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          description: Результат попытки после применения выбора.
        outcome:
          $ref: "#/components/schemas/OutcomeType"
        acceptedChoice:
          $ref: "#/components/schemas/AcceptedChoice"
        consequence:
          $ref: "#/components/schemas/AttemptConsequence"
        revealedNodes:
          type: array
          description: Узлы, раскрытые этим шагом, в порядке появления.
          items:
            $ref: "#/components/schemas/AttemptNode"
        currentNodeId:
          type: string
          description: Узел, на котором прохождение остановилось.
        completedAt:
          type: string
          format: date-time
          description: Заполняется, когда шаг довёл попытку до финала.
...
    ErrorResponse:
      type: object
      description: Единый формат ошибки для всех endpoint.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Стабильный код ошибки. Клиент опирается на код, а не на текст
                сообщения: текст может меняться.
              enum:
                - NOT_FOUND
                - INVALID_REQUEST
                - PAYLOAD_TOO_LARGE
                - SCENARIO_NOT_FOUND
                - UNSUPPORTED_ROLE
                - ATTEMPT_NOT_FOUND
                - ATTEMPT_FORBIDDEN
                - ATTEMPT_ALREADY_COMPLETED
                - STALE_NODE
                - IDEMPOTENCY_KEY_CONFLICT
                - CONCURRENT_TRANSITION
                - NODE_NOT_DECISION
                - CHOICE_NOT_FOUND
                - WEEKLY_TEST_NOT_FOUND
                - WEEKLY_TEST_FORBIDDEN
                - WEEKLY_TEST_INVALID_ANSWERS
                - WEEKLY_TEST_ALREADY_COMPLETED
                - INTERNAL_ERROR
            message:
              type: string
              description: Понятное человеку описание без внутренних подробностей.
            requestId:
              type: string
              description: |
                Идентификатор запроса, совпадает с заголовком `X-Request-ID`.
                Используется для поиска записи в логах.
...